pyproject.toml
[project]
name = "azure-ai-vision-face-api-mcp-server"
version = "1.0.0"
description = "A Face Detection and Recognition MCP Server to allow the embedding of face attribute detection and similar face recognition during Agentic AI workflows."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE.md"]
requires-python = ">=3.10"
dependencies = [
    "azure-ai-vision-face>=1.0.0b2",
    "fastmcp>=2.10.5",
    "openai>=1.97.0",
    "opencv-python>=4.12.0.88",
]
[project.scripts]
azure-face-mcp = "mcpServer:run_server"